backend/requirements.txt
# Django y REST Framework
Django==5.2.1
djangorestframework==3.15.2

# CORS para permitir peticiones desde frontend
django-cors-headers==4.3.1

# Documentación API
drf-yasg==1.21.8

# Base de datos (si usas PostgreSQL en lugar de SQLite)
# psycopg2-binary==2.9.9

# Para manejo de imágenes
Pillow==10.4.0

# Utilidades adicionales
python-decouple==3.8  # Para variables de entorno

# Para desarrollo
django-debug-toolbar==4.4.6  # Solo en desarrollo